# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio Builder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SocialLinks {
  /** @nullable */
  github?: string | null;
  /** @nullable */
  linkedin?: string | null;
  /** @nullable */
  twitter?: string | null;
  /** @nullable */
  website?: string | null;
}

export type PortfolioItemType = typeof PortfolioItemType[keyof typeof PortfolioItemType];


export const PortfolioItemType = {
  experience: 'experience',
  education: 'education',
  project: 'project',
  skill: 'skill',
} as const;

export interface PortfolioItem {
  id: number;
  portfolioId: number;
  type: PortfolioItemType;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  isCurrent?: boolean;
  tags?: string[];
  /** @nullable */
  url?: string | null;
  order?: number;
  createdAt: string;
}

export interface Portfolio {
  id: number;
  userId: string;
  username: string;
  displayName: string;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  theme?: string;
  skills?: string[];
  socialLinks?: SocialLinks;
  items?: PortfolioItem[];
  isPublic?: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface PortfolioInput {
  username?: string;
  displayName?: string;
  tagline?: string;
  bio?: string;
  email?: string;
  location?: string;
  avatarUrl?: string;
  theme?: string;
  skills?: string[];
  socialLinks?: SocialLinks;
  isPublic?: boolean;
}

export type PortfolioItemInputType = typeof PortfolioItemInputType[keyof typeof PortfolioItemInputType];


export const PortfolioItemInputType = {
  experience: 'experience',
  education: 'education',
  project: 'project',
  skill: 'skill',
} as const;

export interface PortfolioItemInput {
  type: PortfolioItemInputType;
  title: string;
  subtitle?: string;
  description?: string;
  startDate?: string;
  endDate?: string;
  isCurrent?: boolean;
  tags?: string[];
  url?: string;
  order?: number;
}

export interface UsernameAvailability {
  available: boolean;
  username: string;
}