# lib/api-zod/src/generated/types/portfolio.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { PortfolioItem } from './portfolioItem';
import type { SocialLinks } from './socialLinks';

export interface Portfolio {
  id: number;
  userId: string;
  username: string;
  displayName: string;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  theme?: string;
  skills?: string[];
  socialLinks?: SocialLinks;
  items?: PortfolioItem[];
  isPublic?: boolean;
  createdAt: string;
  updatedAt: string;
}