# lib/api-zod/src/generated/types/portfolioInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { SocialLinks } from './socialLinks';

export interface PortfolioInput {
  username?: string;
  displayName?: string;
  tagline?: string;
  bio?: string;
  email?: string;
  location?: string;
  avatarUrl?: string;
  theme?: string;
  skills?: string[];
  socialLinks?: SocialLinks;
  isPublic?: boolean;
}