# lib/api-zod/src/generated/types/portfolioItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { PortfolioItemType } from './portfolioItemType';

export interface PortfolioItem {
  id: number;
  portfolioId: number;
  type: PortfolioItemType;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  isCurrent?: boolean;
  tags?: string[];
  /** @nullable */
  url?: string | null;
  order?: number;
  createdAt: string;
}