# lib/api-zod/src/generated/types/portfolioItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { PortfolioItemInputType } from './portfolioItemInputType';

export interface PortfolioItemInput {
  type: PortfolioItemInputType;
  title: string;
  subtitle?: string;
  description?: string;
  startDate?: string;
  endDate?: string;
  isCurrent?: boolean;
  tags?: string[];
  url?: string;
  order?: number;
}