# lib/api-zod/src/generated/types/usernameAvailability.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface UsernameAvailability {
  available: boolean;
  username: string;
}